requirements.txt
asgiref==3.8.1
chardet==5.2.0
Django==5.0.7
djangorestframework==3.15.2
pip==24.0
sqlparse==0.5.0
bs4==0.0.2